move /map to index
